remove start/stop logic
