+++ integration-tests/src/test/java/integration/RestAssuredSetupRule.java
@@ -21,7 +21,7 @@ import com.jayway.restassured.RestAssured;
 import com.jayway.restassured.authentication.AuthenticationScheme;
 import com.jayway.restassured.builder.RequestSpecBuilder;
 import com.jayway.restassured.config.MatcherConfig;
-import integration.util.graylog.GraylogControl;
+import integration.util.graylog.ServerHelper;
 import org.hamcrest.BaseMatcher;
 import org.hamcrest.Description;
 import org.junit.rules.ExternalResource;
@@ -38,9 +38,8 @@ public class RestAssuredSetupRule extends ExternalResource {
 
     @Override
     protected void before() throws Throwable {
-        final GraylogControl graylogController = new GraylogControl();
-        final URL url = graylogController.getUrl();
-        graylogController.startServer();
+        final ServerHelper serverHelper = new ServerHelper();
+        final URL url = serverHelper.getUrl();
         RestAssured.baseURI = url.getProtocol() + "://" + url.getHost();
         RestAssured.port = url.getPort();
         String[] userInfo = url.getUserInfo().split(":");
